Moved meth to func logic to scripts folder
